+++ allensdk/brain_observatory/behavior/behavior_ophys_session.py
@@ -17,7 +17,6 @@ from allensdk.brain_observatory.behavior.dprime import (
 from allensdk.brain_observatory.behavior.dprime import (
     get_hit_rate, get_false_alarm_rate)
 from allensdk.brain_observatory.behavior.image_api import Image, ImageApi
-from allensdk.brain_observatory.running_speed import RunningSpeed
 
 
 class BehaviorOphysSession(ParamsMixin):
@@ -29,7 +28,8 @@ class BehaviorOphysSession(ParamsMixin):
     @classmethod
     def from_lims(cls, ophys_experiment_id: int,
                   eye_tracking_z_threshold: float = 3.0,
-                  eye_tracking_dilation_frames: int = 2) -> "BehaviorOphysSession":
+                  eye_tracking_dilation_frames: int = 2
+                  ) -> "BehaviorOphysSession":
         return cls(api=BehaviorOphysLimsApi(ophys_experiment_id),
                    eye_tracking_z_threshold=eye_tracking_z_threshold,
                    eye_tracking_dilation_frames=eye_tracking_dilation_frames)
@@ -70,9 +70,7 @@ class BehaviorOphysSession(ParamsMixin):
         self._events = None
         self._cell_specimen_table = None
         self._running_speed = None
-        self._running_data_df = None
         self._raw_running_speed = None
-        self._raw_running_data_df = None
         self._stimulus_presentations = None
         self._stimulus_templates = None
         self._licks = None
@@ -199,20 +197,25 @@ class BehaviorOphysSession(ParamsMixin):
         self._cell_specimen_table = value
 
     @property
-    def running_speed(self) -> RunningSpeed:
+    def running_speed(self) -> pd.DataFrame:
         """Running speed of mouse, filtered with 10Hz low pass filter,
-        To get the unfiltered running speed, use `raw_running_speed`.
+
         Returns
         -------
-        allensdk.brain_observatory.running_speed.RunningSpeed
-            NamedTuple with two fields:
-            timestamps : numpy.ndarray
-                Timestamps of running speed data samples
-            values : np.ndarray
-                Running speed of the experimental subject (in cm / s).
+        self._running_speed: pd.DataFrame
+            colums:
+            - timestamps [seconds]
+            - values [cm/s]
+
+        Notes
+        -----
+        To get the unfiltered running speed, use `raw_running_speed`.
         """
         if self._running_speed is None:
-            self._running_speed = self.api.get_running_speed()
+            rs_tuple = self.api.get_running_speed()
+            self._running_speed = pd.DataFrame(
+                    {"timestamps": rs_tuple.timestamps,
+                     "values": rs_tuple.values})
         return self._running_speed
 
     @running_speed.setter
@@ -220,64 +223,30 @@ class BehaviorOphysSession(ParamsMixin):
         self._running_speed = value
 
     @property
-    def raw_running_speed(self) -> RunningSpeed:
-        """Running speed of mouse, unfiltered.
+    def raw_running_speed(self) -> pd.DataFrame:
+        """Running speed of mouse, unfiltered
 
         Returns
         -------
-        allensdk.brain_observatory.running_speed.RunningSpeed
-            NamedTuple with two fields:
-            timestamps : numpy.ndarray
-                Timestamps of running speed data samples
-            values : np.ndarray
-                Running speed of the experimental subject (in cm / s).
+        self._running_speed: pd.DataFrame
+            colums:
+            - timestamps [seconds]
+            - values [cm/s]
+
+        Notes
+        -----
+        To get the filtered running speed, use `running_speed`.
         """
         if self._raw_running_speed is None:
-            self._raw_running_speed = self.api.get_running_speed(lowpass=False)
+            rs_tuple = self.api.get_running_speed(lowpass=False)
+            self._raw_running_speed = pd.DataFrame(
+                    {"timestamps": rs_tuple.timestamps,
+                     "values": rs_tuple.values})
         return self._raw_running_speed
 
     @raw_running_speed.setter
     def raw_running_speed(self, value):
         self._raw_running_speed = value
-
-    @property
-    def running_data_df(self) -> pd.DataFrame:
-        """Get running speed data. By default applies a 10Hz low pass
-        filter to the data. To get the running speed without the filter,
-        use `raw_running_data_df`.
-
-        Returns
-        -------
-        pd.DataFrame
-            Dataframe containing various signals used to compute running
-            speed, and the filtered speed.
-        """
-        if self._running_data_df is None:
-            self._running_data_df = self.api.get_running_data_df(lowpass=True)
-        return self._running_data_df
-
-    @running_data_df.setter
-    def running_data_df(self, value):
-        self._running_data_df = value
-
-    @property
-    def raw_running_data_df(self) -> pd.DataFrame:
-        """Get running speed  data. Does not apply the 10Hz filter.
-
-        Returns
-        -------
-        pd.DataFrame
-            Dataframe containing various signals used to compute running
-            speed, and the unfilteredfiltered speed.
-        """
-        if self._raw_running_data_df is None:
-            self._raw_running_data_df = self.api.get_running_data_df(
-                lowpass=False)
-        return self._raw_running_data_df
-
-    @raw_running_data_df.setter
-    def raw_running_data_df(self, value):
-        self._raw_running_data_df = value
 
     @property
     def stimulus_presentations(self) -> pd.DataFrame:
@@ -447,8 +416,9 @@ class BehaviorOphysSession(ParamsMixin):
         params = {'eye_tracking_dilation_frames', 'eye_tracking_z_threshold'}
 
         if (self._eye_tracking is None) or self.needs_data_refresh(params):
-            self._eye_tracking = self.api.get_eye_tracking(z_threshold=self._eye_tracking_z_threshold,
-                                                           dilation_frames=self._eye_tracking_dilation_frames)
+            self._eye_tracking = self.api.get_eye_tracking(
+                    z_threshold=self._eye_tracking_z_threshold,
+                    dilation_frames=self._eye_tracking_dilation_frames)
             self.clear_updated_params(params)
 
         return self._eye_tracking
@@ -459,8 +429,15 @@ class BehaviorOphysSession(ParamsMixin):
 
     @property
     def eye_tracking_rig_geometry(self) -> dict:
-        """Get the eye tracking rig geometry associated with an ophys experiment"""
+        """Get the eye tracking rig geometry
+        associated with an ophys experiment"""
         return self.api.get_eye_tracking_rig_geometry()
+
+    @property
+    def eye_gaze_mapping_file_path(self) -> str:
+        """Get h5 filepath containing eye gaze behavior of
+        the experiment's subject"""
+        return self.api.get_eye_gaze_mapping_file_path()
 
     def cache_clear(self) -> None:
         """Convenience method to clear the api cache, if applicable."""
@@ -471,43 +448,9 @@ class BehaviorOphysSession(ParamsMixin):
                 "Attempted to clear API cache, but method `cache_clear`"
                 f" does not exist on {self.api.__class__.__name__}")
 
-    def get_roi_masks(self, cell_specimen_ids=None) -> xr.DataArray:
-        """ Obtains boolean masks indicating the location of one or
-        more cell's ROIs in this session.
-
-        Parameters
-        ----------
-        cell_specimen_ids : array-like of int, optional
-            ROI masks for these cell specimens will be returned. The default
-            behavior is to return masks for all cell specimens.
-
-        Returns
-        -------
-        result : xr.DataArray
-            dimensions are:
-                - cell_specimen_id : which cell's roi is described by the mask
-                - row : index within the underlying image
-                - column : index within the image
-            values are 1 where an ROI was present, otherwise 0.
-        """
-        cell_specimen_table = self.cell_specimen_table
-
-        if cell_specimen_ids is None:
-            cell_specimen_ids = cell_specimen_table.index.values
-        elif (isinstance(cell_specimen_ids, int)
-              or np.issubdtype(type(cell_specimen_ids), np.integer)):
-            cell_specimen_ids = np.array([int(cell_specimen_ids)])
-        else:
-            cell_specimen_ids = np.array(cell_specimen_ids)
-
-        cell_roi_ids = cell_specimen_table.loc[cell_specimen_ids,
-                                               "cell_roi_id"].values
-        result = self.api.get_roi_masks_by_cell_roi_id(cell_roi_ids)
-        if "cell_roi_id" in result.dims:
-            result = result.rename({"cell_roi_id": "cell_specimen_id"})
-            result.coords["cell_specimen_id"] = cell_specimen_ids
-
-        return result
+    @property
+    def roi_masks(self) -> pd.DataFrame:
+        return self.cell_specimen_table[['cell_roi_id', 'roi_mask']]
 
     @legacy('Consider using "dff_traces" instead.')
     def get_dff_traces(self, cell_specimen_ids=None):
@@ -515,8 +458,10 @@ class BehaviorOphysSession(ParamsMixin):
         if cell_specimen_ids is None:
             cell_specimen_ids = self.get_cell_specimen_ids()
 
-        csid_table = self.cell_specimen_table.reset_index()[['cell_specimen_id']]
-        csid_subtable = csid_table[csid_table['cell_specimen_id'].isin(cell_specimen_ids)].set_index('cell_specimen_id')
+        csid_table = \
+            self.cell_specimen_table.reset_index()[['cell_specimen_id']]
+        csid_subtable = csid_table[csid_table['cell_specimen_id'].isin(
+            cell_specimen_ids)].set_index('cell_specimen_id')
         dff_table = csid_subtable.join(self.dff_traces, how='left')
         dff_traces = np.vstack(dff_table['dff'].values)
         timestamps = self.ophys_timestamps
@@ -526,14 +471,17 @@ class BehaviorOphysSession(ParamsMixin):
 
     @legacy()
     def get_cell_specimen_indices(self, cell_specimen_ids):
-        return [self.cell_specimen_table.index.get_loc(csid) for csid in cell_specimen_ids]
+        return [self.cell_specimen_table.index.get_loc(csid)
+                for csid in cell_specimen_ids]
 
-    @legacy('Consider using "cell_specimen_table[\'cell_specimen_id\']" instead.')
+    @legacy("Consider using cell_specimen_table['cell_specimen_id'] instead.")
     def get_cell_specimen_ids(self):
         cell_specimen_ids = self.cell_specimen_table.index.values
 
-        if np.isnan(cell_specimen_ids.astype(float)).sum() == len(self.cell_specimen_table):
-            raise ValueError(f'cell_specimen_id values not assigned for {self.ophys_experiment_id}')
+        if np.isnan(cell_specimen_ids.astype(float)).sum() == \
+                len(self.cell_specimen_table):
+            raise ValueError("cell_specimen_id values not assigned "
+                             f"for {self.ophys_experiment_id}")
         return cell_specimen_ids
 
     def deserialize_image(self, sitk_image):
@@ -550,7 +498,8 @@ class BehaviorOphysSession(ParamsMixin):
         return img
 
     def get_max_projection(self):
-        """ Returns an image whose values are the maximum obtained values at each pixel of the ophys movie over time.
+        """ Returns an image whose values are the maximum obtained
+        values at each pixel of the ophys movie over time.
 
         Returns
         ----------
@@ -560,7 +509,8 @@ class BehaviorOphysSession(ParamsMixin):
         return self.deserialize_image(self.api.get_max_projection())
 
     def get_average_projection(self):
-        """ Returns an image whose values are the average obtained values at each pixel of the ophys movie over time.
+        """ Returns an image whose values are the average obtained
+        values at each pixel of the ophys movie over time.
 
         Returns
         ----------
@@ -570,12 +520,14 @@ class BehaviorOphysSession(ParamsMixin):
         return self.deserialize_image(self.api.get_average_projection())
 
     def get_segmentation_mask_image(self):
-        """ Returns an image with value 1 if the pixel was included in an ROI, and 0 otherwise
+        """ Returns an image with value 1 if the pixel was included
+        in an ROI, and 0 otherwise
 
         Returns
         ----------
         allensdk.brain_observatory.behavior.image_api.Image:
             array-like interface to segmentation_mask image data and metadata
+
         """
         masks = self.api.get_roi_masks_by_cell_roi_id()
         mask_image_data = masks.any(dim='cell_roi_id').astype(int)
@@ -589,44 +541,72 @@ class BehaviorOphysSession(ParamsMixin):
     def get_reward_rate(self):
         response_latency_list = []
         for _, t in self.trials.iterrows():
-            valid_response_licks = [l for l in t.lick_times if l - t.change_time > self.task_parameters['response_window_sec'][0]]
-            response_latency = float('inf') if len(valid_response_licks) == 0 else valid_response_licks[0] - t.change_time
+            valid_response_licks = \
+                    [l for l in t.lick_times
+                     if l - t.change_time >
+                        self.task_parameters['response_window_sec'][0]]
+            response_latency = (
+                    float('inf')
+                    if len(valid_response_licks) == 0
+                    else valid_response_licks[0] - t.change_time)
             response_latency_list.append(response_latency)
-        reward_rate = calculate_reward_rate(response_latency=response_latency_list, starttime=self.trials.start_time.values)
+        reward_rate = calculate_reward_rate(
+                response_latency=response_latency_list,
+                starttime=self.trials.start_time.values)
         reward_rate[np.isinf(reward_rate)] = float('nan')
         return reward_rate
 
     def get_rolling_performance_df(self):
-
         # Indices to build trial metrics dataframe:
         trials_index = self.trials.index
-        not_aborted_index = self.trials[np.logical_not(self.trials.aborted)].index
+        not_aborted_index = \
+            self.trials[np.logical_not(self.trials.aborted)].index
 
         # Initialize dataframe:
         performance_metrics_df = pd.DataFrame(index=trials_index)
 
         # Reward rate:
-        performance_metrics_df['reward_rate'] = pd.Series(self.get_reward_rate(), index=self.trials.index)
+        performance_metrics_df['reward_rate'] = \
+            pd.Series(self.get_reward_rate(), index=self.trials.index)
 
         # Hit rate raw:
-        hit_rate_raw = get_hit_rate(hit=self.trials.hit, miss=self.trials.miss, aborted=self.trials.aborted)
-        performance_metrics_df['hit_rate_raw'] = pd.Series(hit_rate_raw, index=not_aborted_index)
+        hit_rate_raw = get_hit_rate(
+            hit=self.trials.hit,
+            miss=self.trials.miss,
+            aborted=self.trials.aborted)
+        performance_metrics_df['hit_rate_raw'] = \
+            pd.Series(hit_rate_raw, index=not_aborted_index)
 
         # Hit rate with trial count correction:
-        hit_rate = get_trial_count_corrected_hit_rate(hit=self.trials.hit, miss=self.trials.miss, aborted=self.trials.aborted)
-        performance_metrics_df['hit_rate'] = pd.Series(hit_rate, index=not_aborted_index)
+        hit_rate = get_trial_count_corrected_hit_rate(
+                hit=self.trials.hit,
+                miss=self.trials.miss,
+                aborted=self.trials.aborted)
+        performance_metrics_df['hit_rate'] = \
+            pd.Series(hit_rate, index=not_aborted_index)
 
         # False-alarm rate raw:
-        false_alarm_rate_raw = get_false_alarm_rate(false_alarm=self.trials.false_alarm, correct_reject=self.trials.correct_reject, aborted=self.trials.aborted)
-        performance_metrics_df['false_alarm_rate_raw'] = pd.Series(false_alarm_rate_raw, index=not_aborted_index)
+        false_alarm_rate_raw = \
+            get_false_alarm_rate(
+                    false_alarm=self.trials.false_alarm,
+                    correct_reject=self.trials.correct_reject,
+                    aborted=self.trials.aborted)
+        performance_metrics_df['false_alarm_rate_raw'] = \
+            pd.Series(false_alarm_rate_raw, index=not_aborted_index)
 
         # False-alarm rate with trial count correction:
-        false_alarm_rate = get_trial_count_corrected_false_alarm_rate(false_alarm=self.trials.false_alarm, correct_reject=self.trials.correct_reject, aborted=self.trials.aborted)
-        performance_metrics_df['false_alarm_rate'] = pd.Series(false_alarm_rate, index=not_aborted_index)
+        false_alarm_rate = \
+            get_trial_count_corrected_false_alarm_rate(
+                    false_alarm=self.trials.false_alarm,
+                    correct_reject=self.trials.correct_reject,
+                    aborted=self.trials.aborted)
+        performance_metrics_df['false_alarm_rate'] = \
+            pd.Series(false_alarm_rate, index=not_aborted_index)
 
         # Rolling-dprime:
         rolling_dprime = get_rolling_dprime(hit_rate, false_alarm_rate)
-        performance_metrics_df['rolling_dprime'] = pd.Series(rolling_dprime, index=not_aborted_index)
+        performance_metrics_df['rolling_dprime'] = \
+            pd.Series(rolling_dprime, index=not_aborted_index)
 
         return performance_metrics_df
 
@@ -637,27 +617,44 @@ class BehaviorOphysSession(ParamsMixin):
         performance_metrics['catch_trial_count'] = self.trials.catch.sum()
         performance_metrics['hit_trial_count'] = self.trials.hit.sum()
         performance_metrics['miss_trial_count'] = self.trials.miss.sum()
-        performance_metrics['false_alarm_trial_count'] = self.trials.false_alarm.sum()
-        performance_metrics['correct_reject_trial_count'] = self.trials.correct_reject.sum()
-        performance_metrics['auto_rewarded_trial_count'] = self.trials.auto_rewarded.sum()
-        performance_metrics['rewarded_trial_count'] = self.trials.reward_time.apply(lambda x: not np.isnan(x)).sum()
+        performance_metrics['false_alarm_trial_count'] = \
+            self.trials.false_alarm.sum()
+        performance_metrics['correct_reject_trial_count'] = \
+            self.trials.correct_reject.sum()
+        performance_metrics['auto_rewarded_trial_count'] = \
+            self.trials.auto_rewarded.sum()
+        performance_metrics['rewarded_trial_count'] = \
+            self.trials.reward_time.apply(lambda x: not np.isnan(x)).sum()
         performance_metrics['total_reward_count'] = len(self.rewards)
         performance_metrics['total_reward_volume'] = self.rewards.volume.sum()
 
-        rolling_performance_df = self.get_rolling_performance_df()
-        engaged_trial_mask = (rolling_performance_df['reward_rate'] > engaged_trial_reward_rate_threshold)
-        performance_metrics['maximum_reward_rate'] = np.nanmax(rolling_performance_df['reward_rate'].values)
+        rpdf = self.get_rpdf()
+        engaged_trial_mask = (
+                rpdf['reward_rate'] >
+                engaged_trial_reward_rate_threshold)
+        performance_metrics['maximum_reward_rate'] = \
+            np.nanmax(rpdf['reward_rate'].values)
         performance_metrics['engaged_trial_count'] = (engaged_trial_mask).sum()
-        performance_metrics['mean_hit_rate'] = rolling_performance_df['hit_rate'].mean()
-        performance_metrics['mean_hit_rate_uncorrected'] = rolling_performance_df['hit_rate_raw'].mean()
-        performance_metrics['mean_hit_rate_engaged'] = rolling_performance_df['hit_rate'][engaged_trial_mask].mean()
-        performance_metrics['mean_false_alarm_rate'] = rolling_performance_df['false_alarm_rate'].mean()
-        performance_metrics['mean_false_alarm_rate_uncorrected'] = rolling_performance_df['false_alarm_rate_raw'].mean()
-        performance_metrics['mean_false_alarm_rate_engaged'] = rolling_performance_df['false_alarm_rate'][engaged_trial_mask].mean()
-        performance_metrics['mean_dprime'] = rolling_performance_df['rolling_dprime'].mean()
-        performance_metrics['mean_dprime_engaged'] = rolling_performance_df['rolling_dprime'][engaged_trial_mask].mean()
-        performance_metrics['max_dprime'] = rolling_performance_df['rolling_dprime'].max()
-        performance_metrics['max_dprime_engaged'] = rolling_performance_df['rolling_dprime'][engaged_trial_mask].max()
+        performance_metrics['mean_hit_rate'] = \
+            rpdf['hit_rate'].mean()
+        performance_metrics['mean_hit_rate_uncorrected'] = \
+            rpdf['hit_rate_raw'].mean()
+        performance_metrics['mean_hit_rate_engaged'] = \
+            rpdf['hit_rate'][engaged_trial_mask].mean()
+        performance_metrics['mean_false_alarm_rate'] = \
+            rpdf['false_alarm_rate'].mean()
+        performance_metrics['mean_false_alarm_rate_uncorrected'] = \
+            rpdf['false_alarm_rate_raw'].mean()
+        performance_metrics['mean_false_alarm_rate_engaged'] = \
+            rpdf['false_alarm_rate'][engaged_trial_mask].mean()
+        performance_metrics['mean_dprime'] = \
+            rpdf['rolling_dprime'].mean()
+        performance_metrics['mean_dprime_engaged'] = \
+            rpdf['rolling_dprime'][engaged_trial_mask].mean()
+        performance_metrics['max_dprime'] = \
+            rpdf['rolling_dprime'].max()
+        performance_metrics['max_dprime_engaged'] = \
+            rpdf['rolling_dprime'][engaged_trial_mask].max()
 
         return performance_metrics