+++ allensdk/test/brain_observatory/behavior/test_behavior_ophys_session.py
@@ -169,44 +169,6 @@ def test_trial_response_window_bounds_reward(ophys_experiment_id):
                 assert lick_times[0] < reward_time
 
 
-@pytest.mark.parametrize("cell_specimen_ids,expected", [
-    [
-        10,
-        np.array([
-            [1, 0, 0, 0, 0],
-            [0, 0, 0, 0, 0],
-            [0, 0, 0, 0, 0],
-            [0, 0, 0, 0, 0],
-            [0, 0, 0, 0, 0]
-        ])
-    ],
-    [
-        [11, 10],
-        np.array([
-            [
-                [0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0],
-                [0, 1, 0, 0, 0],
-                [0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0]
-            ],
-            [
-                [1, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0]
-            ]
-        ])
-    ]
-])
-# cell_specimen_table_api fixture from allensdk.test.brain_observatory.conftest
-def test_get_roi_masks(cell_specimen_ids, expected, cell_specimen_table_api):
-    ssn = BehaviorOphysSession(api=cell_specimen_table_api)
-    obtained = ssn.get_roi_masks(cell_specimen_ids)
-    assert np.allclose(expected, obtained.values)
-
-
 @pytest.mark.parametrize("dilation_frames, z_threshold, eye_tracking_start_value", [
     (5, 9, None),
     (1, 2, None),
@@ -229,23 +191,3 @@ def test_eye_tracking(dilation_frames, z_threshold, eye_tracking_start_value):
         assert obtained.equals(pd.DataFrame([1, 2, 3]))
         assert session.api.get_eye_tracking.called_with(z_threshold=z_threshold,
                                                         dilation_frames=dilation_frames)
-
-
-# @pytest.mark.requires_bamboo
-def test_event_detection():
-    ophys_experiment_id = 789359614
-    session = BehaviorOphysSession.from_lims(ophys_experiment_id=ophys_experiment_id)
-    events = session.events
-
-    assert len(events) > 0
-
-    expected_columns = ['events', 'filtered_events', 'lambda', 'noise_std', 'cell_roi_id']
-    assert len(events.columns) == len(expected_columns)
-    # Assert they contain the same columns
-    assert len(set(expected_columns).intersection(events.columns)) == len(expected_columns)
-
-    assert events.index.name == 'cell_specimen_id'
-
-    # All events are the same length
-    event_length = len(set([len(x) for x in events['events']]))
-    assert event_length == 1