+++ allensdk/test/brain_observatory/behavior/test_write_nwb_behavior_ophys.py
@@ -9,8 +9,7 @@ import pytest
 import allensdk.brain_observatory.nwb as nwb
 from allensdk.brain_observatory.behavior.session_apis.data_io import (
     BehaviorOphysNwbApi)
-from allensdk.test.brain_observatory.behavior.test_eye_tracking_processing import create_preload_eye_tracking_df, \
-    create_refined_eye_tracking_df
+from allensdk.test.brain_observatory.behavior.test_eye_tracking_processing import create_refined_eye_tracking_df  # noqa: E501
 
 
 @pytest.fixture
@@ -54,50 +53,30 @@ def eye_gaze_data():
 @pytest.fixture
 def rig_geometry():
     """Returns mock rig geometry data"""
-    return {"monitor_position_mm": [1., 2., 3.],
-     "monitor_rotation_deg": [4., 5., 6.],
-     "camera_position_mm": [7., 8., 9.],
-     "camera_rotation_deg": [10., 11., 12.],
-     "led_position": [13., 14., 15.],
-     "equipment": "test_rig"}
+    return {
+            "monitor_position_mm": [1., 2., 3.],
+            "monitor_rotation_deg": [4., 5., 6.],
+            "camera_position_mm": [7., 8., 9.],
+            "camera_rotation_deg": [10., 11., 12.],
+            "led_position": [13., 14., 15.],
+            "equipment": "test_rig"}
 
 
 @pytest.fixture
 def eye_tracking_data():
     return create_refined_eye_tracking_df(
         np.array([[0.1, 12 * np.pi, 72 * np.pi, 196 * np.pi, False,
-                   1., 2., 3., 4., 5., 6., 7., 8., 9., 10., 11., 12., 13., 14., 15.],
+                   1., 2., 3., 4., 5., 6., 7., 8., 9., 10., 11., 12.,
+                   13., 14., 15.],
                   [0.2, 20 * np.pi, 90 * np.pi, 225 * np.pi, False,
-                   2., 3., 4., 5., 6., 7., 8., 9., 10., 11., 12., 13., 14., 15., 16.]])
+                   2., 3., 4., 5., 6., 7., 8., 9., 10., 11., 12., 13.,
+                   14., 15., 16.]])
     )
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
-def test_add_running_data_dfs_to_nwbfile(nwbfile, running_data_df, roundtrip, roundtripper):
-    # Just make it different from running_data_df
-    running_data_df_unfiltered = running_data_df.copy()
-    running_data_df_unfiltered['speed'] = running_data_df['speed'] * 2
-
-    unit_dict = {'v_sig': 'V', 'v_in': 'V', 'speed': 'cm/s', 'timestamps': 's', 'dx': 'cm'}
-    nwbfile = nwb.add_running_data_dfs_to_nwbfile(
-        nwbfile,
-        running_data_df=running_data_df,
-        running_data_df_unfiltered=running_data_df_unfiltered,
-        unit_dict=unit_dict)
-
-    if roundtrip:
-        obt = roundtripper(nwbfile, BehaviorOphysNwbApi)
-    else:
-        obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
-
-    pd.testing.assert_frame_equal(
-        running_data_df, obt.get_running_data_df(lowpass=True))
-    pd.testing.assert_frame_equal(
-        running_data_df_unfiltered, obt.get_running_data_df(lowpass=False))
-
-
-@pytest.mark.parametrize('roundtrip', [True, False])
-def test_add_stimulus_templates(nwbfile, stimulus_templates, roundtrip, roundtripper):
+def test_add_stimulus_templates(nwbfile, stimulus_templates,
+                                roundtrip, roundtripper):
     for key, val in stimulus_templates.items():
         nwb.add_stimulus_template(nwbfile, val, key)
 
@@ -107,12 +86,17 @@ def test_add_stimulus_templates(nwbfile, stimulus_templates, roundtrip, roundtri
         obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
 
     stimulus_templates_obt = obt.get_stimulus_templates()
-    for key in set(stimulus_templates.keys()).union(set(stimulus_templates_obt.keys())):
-        np.testing.assert_array_almost_equal(stimulus_templates[key], stimulus_templates_obt[key])
+    uset = set(stimulus_templates.keys()).union(
+            set(stimulus_templates_obt.keys()))
+    for key in uset:
+        np.testing.assert_array_almost_equal(stimulus_templates[key],
+                                             stimulus_templates_obt[key])
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
-def test_add_stimulus_presentations(nwbfile, stimulus_presentations_behavior, stimulus_timestamps, roundtrip, roundtripper, stimulus_templates):
+def test_add_stimulus_presentations(nwbfile, stimulus_presentations_behavior,
+                                    stimulus_timestamps, roundtrip,
+                                    roundtripper, stimulus_templates):
     nwb.add_stimulus_timestamps(nwbfile, stimulus_timestamps)
     nwb.add_stimulus_presentations(nwbfile, stimulus_presentations_behavior)
     for key, val in stimulus_templates.items():
@@ -120,7 +104,9 @@ def test_add_stimulus_presentations(nwbfile, stimulus_presentations_behavior, st
 
         # Add index for this template to NWB in-memory object:
         nwb_template = nwbfile.stimulus_template[key]
-        curr_stimulus_index = stimulus_presentations_behavior[stimulus_presentations_behavior['image_set'] == nwb_template.name]
+        compare = (stimulus_presentations_behavior['image_set'] ==
+                   nwb_template.name)
+        curr_stimulus_index = stimulus_presentations_behavior[compare]
         nwb.add_stimulus_index(nwbfile, curr_stimulus_index, nwb_template)
 
     if roundtrip:
@@ -128,11 +114,14 @@ def test_add_stimulus_presentations(nwbfile, stimulus_presentations_behavior, st
     else:
         obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
 
-    pd.testing.assert_frame_equal(stimulus_presentations_behavior, obt.get_stimulus_presentations(), check_dtype=False)
+    pd.testing.assert_frame_equal(stimulus_presentations_behavior,
+                                  obt.get_stimulus_presentations(),
+                                  check_dtype=False)
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
-def test_add_stimulus_timestamps(nwbfile, stimulus_timestamps, roundtrip, roundtripper):
+def test_add_stimulus_timestamps(nwbfile, stimulus_timestamps,
+                                 roundtrip, roundtripper):
 
     nwb.add_stimulus_timestamps(nwbfile, stimulus_timestamps)
 
@@ -141,7 +130,8 @@ def test_add_stimulus_timestamps(nwbfile, stimulus_timestamps, roundtrip, roundt
     else:
         obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
 
-    np.testing.assert_array_almost_equal(stimulus_timestamps, obt.get_stimulus_timestamps())
+    np.testing.assert_array_almost_equal(stimulus_timestamps,
+                                         obt.get_stimulus_timestamps())
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
@@ -180,11 +170,13 @@ def test_add_rewards(nwbfile, roundtrip, roundtripper, rewards):
     else:
         obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
 
-    pd.testing.assert_frame_equal(rewards, obt.get_rewards(), check_dtype=False)
+    pd.testing.assert_frame_equal(rewards, obt.get_rewards(),
+                                  check_dtype=False)
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
-def test_add_max_projection(nwbfile, roundtrip, roundtripper, max_projection, image_api):
+def test_add_max_projection(nwbfile, roundtrip, roundtripper,
+                            max_projection, image_api):
 
     nwb.add_max_projection(nwbfile, max_projection)
 
@@ -193,11 +185,13 @@ def test_add_max_projection(nwbfile, roundtrip, roundtripper, max_projection, im
     else:
         obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
 
-    assert image_api.deserialize(max_projection) == image_api.deserialize(obt.get_max_projection())
+    assert image_api.deserialize(max_projection) == \
+        image_api.deserialize(obt.get_max_projection())
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
-def test_add_average_image(nwbfile, roundtrip, roundtripper, average_image, image_api):
+def test_add_average_image(nwbfile, roundtrip, roundtripper, average_image,
+                           image_api):
 
     nwb.add_average_image(nwbfile, average_image)
 
@@ -206,11 +200,13 @@ def test_add_average_image(nwbfile, roundtrip, roundtripper, average_image, imag
     else:
         obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
 
-    assert image_api.deserialize(average_image) == image_api.deserialize(obt.get_average_projection())
+    assert image_api.deserialize(average_image) == \
+        image_api.deserialize(obt.get_average_projection())
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
-def test_segmentation_mask_image(nwbfile, roundtrip, roundtripper, segmentation_mask_image, image_api):
+def test_segmentation_mask_image(nwbfile, roundtrip, roundtripper,
+                                 segmentation_mask_image, image_api):
 
     nwb.add_segmentation_mask_image(nwbfile, segmentation_mask_image)
 
@@ -219,7 +215,8 @@ def test_segmentation_mask_image(nwbfile, roundtrip, roundtripper, segmentation_
     else:
         obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
 
-    assert image_api.deserialize(segmentation_mask_image) == image_api.deserialize(obt.get_segmentation_mask_image())
+    assert image_api.deserialize(segmentation_mask_image) == \
+        image_api.deserialize(obt.get_segmentation_mask_image())
 
 
 @pytest.mark.parametrize('test_partial_metadata', [True, False])
@@ -258,7 +255,8 @@ def test_add_partial_metadata(test_partial_metadata, roundtrip, roundtripper,
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
-def test_add_task_parameters(nwbfile, roundtrip, roundtripper, task_parameters):
+def test_add_task_parameters(nwbfile, roundtrip,
+                             roundtripper, task_parameters):
 
     nwb.add_task_parameters(nwbfile, task_parameters)
 
@@ -282,37 +280,49 @@ def test_add_task_parameters(nwbfile, roundtrip, roundtripper, task_parameters):
 
 @pytest.mark.parametrize('roundtrip', [True, False])
 @pytest.mark.parametrize("filter_invalid_rois", [True, False])
-def test_get_cell_specimen_table(nwbfile, roundtrip, filter_invalid_rois, valid_roi_ids, roundtripper, cell_specimen_table, metadata, ophys_timestamps):
+def test_get_cell_specimen_table(nwbfile, roundtrip, filter_invalid_rois,
+                                 valid_roi_ids, roundtripper,
+                                 cell_specimen_table, metadata,
+                                 ophys_timestamps):
 
     nwb.add_metadata(nwbfile, metadata, behavior_only=False)
     nwb.add_cell_specimen_table(nwbfile, cell_specimen_table, metadata)
 
     if roundtrip:
-        obt = roundtripper(nwbfile, BehaviorOphysNwbApi, filter_invalid_rois=filter_invalid_rois)
+        obt = roundtripper(nwbfile, BehaviorOphysNwbApi,
+                           filter_invalid_rois=filter_invalid_rois)
     else:
-        obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile, filter_invalid_rois=filter_invalid_rois)
+        obt = BehaviorOphysNwbApi.from_nwbfile(
+                nwbfile, filter_invalid_rois=filter_invalid_rois)
 
     if filter_invalid_rois:
-        cell_specimen_table = cell_specimen_table[cell_specimen_table["cell_roi_id"].isin(valid_roi_ids)]
+        cell_specimen_table = \
+                cell_specimen_table[
+                        cell_specimen_table["cell_roi_id"].isin(
+                            valid_roi_ids)]
 
-    print(cell_specimen_table)
-    print(obt.get_cell_specimen_table())
-
-    pd.testing.assert_frame_equal(cell_specimen_table, obt.get_cell_specimen_table(), check_dtype=False)
+    pd.testing.assert_frame_equal(
+            cell_specimen_table,
+            obt.get_cell_specimen_table(),
+            check_dtype=False)
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
 @pytest.mark.parametrize("filter_invalid_rois", [True, False])
-def test_get_dff_traces(nwbfile, roundtrip, filter_invalid_rois, valid_roi_ids, roundtripper, dff_traces, cell_specimen_table, metadata, ophys_timestamps):
+def test_get_dff_traces(nwbfile, roundtrip, filter_invalid_rois, valid_roi_ids,
+                        roundtripper, dff_traces, cell_specimen_table,
+                        metadata, ophys_timestamps):
 
     nwb.add_metadata(nwbfile, metadata, behavior_only=False)
     nwb.add_cell_specimen_table(nwbfile, cell_specimen_table, metadata)
     nwb.add_dff_traces(nwbfile, dff_traces, ophys_timestamps)
 
     if roundtrip:
-        obt = roundtripper(nwbfile, BehaviorOphysNwbApi, filter_invalid_rois=filter_invalid_rois)
+        obt = roundtripper(nwbfile, BehaviorOphysNwbApi,
+                           filter_invalid_rois=filter_invalid_rois)
     else:
-        obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile, filter_invalid_rois=filter_invalid_rois)
+        obt = BehaviorOphysNwbApi.from_nwbfile(
+                nwbfile, filter_invalid_rois=filter_invalid_rois)
 
     if filter_invalid_rois:
         dff_traces = dff_traces[dff_traces["cell_roi_id"].isin(valid_roi_ids)]
@@ -321,21 +331,26 @@ def test_get_dff_traces(nwbfile, roundtrip, filter_invalid_rois, valid_roi_ids, 
 
     print(obt.get_dff_traces())
 
-    pd.testing.assert_frame_equal(dff_traces, obt.get_dff_traces(), check_dtype=False)
+    pd.testing.assert_frame_equal(
+            dff_traces, obt.get_dff_traces(), check_dtype=False)
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
 @pytest.mark.parametrize("filter_invalid_rois", [True, False])
-def test_get_corrected_fluorescence_traces(nwbfile, roundtrip, filter_invalid_rois, valid_roi_ids, roundtripper, dff_traces, corrected_fluorescence_traces, cell_specimen_table, metadata, ophys_timestamps):
+def test_get_corrected_fluorescence_traces(
+        nwbfile, roundtrip, filter_invalid_rois, valid_roi_ids, roundtripper,
+        dff_traces, corrected_fluorescence_traces, cell_specimen_table,
+        metadata, ophys_timestamps):
 
     nwb.add_metadata(nwbfile, metadata, behavior_only=False)
     nwb.add_cell_specimen_table(nwbfile, cell_specimen_table, metadata)
     nwb.add_dff_traces(nwbfile, dff_traces, ophys_timestamps)
-    nwb.add_corrected_fluorescence_traces(nwbfile, corrected_fluorescence_traces)
+    nwb.add_corrected_fluorescence_traces(nwbfile,
+                                          corrected_fluorescence_traces)
 
     if roundtrip:
         obt = roundtripper(nwbfile, BehaviorOphysNwbApi,
-                          filter_invalid_rois=filter_invalid_rois)
+                           filter_invalid_rois=filter_invalid_rois)
     else:
         obt = BehaviorOphysNwbApi.from_nwbfile(
             nwbfile, filter_invalid_rois=filter_invalid_rois)
@@ -353,7 +368,9 @@ def test_get_corrected_fluorescence_traces(nwbfile, roundtrip, filter_invalid_ro
 
 
 @pytest.mark.parametrize('roundtrip', [True, False])
-def test_get_motion_correction(nwbfile, roundtrip, roundtripper, motion_correction, ophys_timestamps, metadata, cell_specimen_table, dff_traces):
+def test_get_motion_correction(nwbfile, roundtrip, roundtripper,
+                               motion_correction, ophys_timestamps,
+                               metadata, cell_specimen_table, dff_traces):
 
     nwb.add_metadata(nwbfile, metadata, behavior_only=False)
     nwb.add_cell_specimen_table(nwbfile, cell_specimen_table, metadata)
@@ -365,7 +382,11 @@ def test_get_motion_correction(nwbfile, roundtrip, roundtripper, motion_correcti
     else:
         obt = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
 
-    pd.testing.assert_frame_equal(motion_correction, obt.get_motion_correction(), check_dtype=False)
+    pd.testing.assert_frame_equal(
+            motion_correction,
+            obt.get_motion_correction(),
+            check_dtype=False)
+
 
 @pytest.mark.parametrize("roundtrip", [True, False])
 @pytest.mark.parametrize("expected", [
@@ -395,8 +416,9 @@ def test_add_eye_tracking_rig_geometry_data_to_nwbfile(nwbfile, roundtripper,
 
 
 @pytest.mark.parametrize("roundtrip", [True, False])
-def test_add_eye_tracking_data_to_nwbfile(tmp_path, nwbfile, eye_tracking_data, rig_geometry,
-                                          roundtripper, roundtrip):
+def test_add_eye_tracking_data_to_nwbfile(
+        tmp_path, nwbfile, eye_tracking_data,
+        rig_geometry, roundtripper, roundtrip):
     api = BehaviorOphysNwbApi.from_nwbfile(nwbfile)
     nwbfile = api.add_eye_tracking_data_to_nwb(
         nwbfile=nwbfile,